+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2020-10-??
+### Added
+- Guest Mode: Play without logging into your Google account. You will be unable to make any suggestions.
+
+### Changed
+- The "Login with Google" button has been swapped with a more visible button.
+
 ## [1.5.2] - 2020-10-06
 ### Fixed
 - You can no longer suggest elements named "Your Element".